% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yart.R
\docType{package}
\name{yart}
\alias{yart}
\alias{yart-package}
\title{LaTeX Report format template YART}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{A boolean toggle to select whether intermediate
LaTeX files are to be kept, defaults to \code{FALSE}}
}
\value{
R Markdown output format to pass to
\code{\link[rmarkdown:render]{render}}
}
\description{
\pkg{yart} provides a template suitable for reports such as academic papers (thesis). It can be used with (or without) \href{http://www.rstudio.com/}{RStudio}.
}
\details{
#' Note, at this point `yart` is in active development and should be considered alpha. If you experience any problems,
please \href{https://github.com/sebastiansauer/yart/issues}{open an issue} with a reproducible example of the encountered problem.
}
\section{System requirements}{

To use \pkg{yart} you need to make sure the following software is installed on your computer:

   \itemize{
       \item{\href{http://www.rstudio.com/}{RStudio} (>= 0.98.932) or if you don't use RStudio
           , you need to install \href{http://johnmacfarlane.net/pandoc/}{pandoc} using the
           \href{https://github.com/rstudio/rmarkdown/blob/master/PANDOC.md}{instructions for your operating system}}
       \item{A \href{http://de.wikipedia.org/wiki/TeX}{TeX} distribution (2013 or later; e.g., \href{http://miktex.org/}{MikTeX}
           for Windows, \href{https://tug.org/mactex/}{MacTeX} for Mac, obviously, or \href{http://www.tug.org/texlive/}{TeX Live}
           for Linux)}
   }

   \emph{Windows} users should use MikTex if possible. Currently, pandoc and the Windows version of Tex Live
   \href{https://github.com/rstudio/rmarkdown/issues/6}{don't seem to like each other}. Make sure you install the
   \emph{complete}---not the basic---version.

   \emph{Ubuntu 14.04} users need a few additional TeX packages for the document class
   \href{http://www.ctan.org/pkg/apa6}{apa6} to work:

   \code{sudo apt-get install texlive texlive-publishers texlive-fonts-extra texlive-latex-extra texlive-humanities lmodern}
}

\section{Authors}{

   Sebastian Sauer (sebastian.sauer at posteo.de)
}

\section{Maintainer}{

   Sebastian Sauer (sebastian.sauer at posteo.de)
}

\section{Report features}{

Various aspects of the report can be customized by setting the following
variables in the document metadata:

\describe{
  \item{\code{title}}{Title of the report, colon is permitted}
  \item{\code{subtitle}}{Subtitle}
  \item{\code{author}}{Author}
  \item{\code{date}}{Date}
  \item{\code{logo}}{logo (image) file, png format preferred}
  \item{\code{field}}{Field such as "Psychology", "Marketing", "Consumer Contact"}
  \item{\code{lang}}{LaTeX language support, eg., "de-De"}
  \item{\code{lof}}{List of Figures, Boolean: TRUE or FALSE}
  \item{\code{lot}}{List of Tables, Boolean: TRUE or FALSE}
  \item{\code{toc}}{Table of Contents, Boolean: TRUE or FALSE}
  \item{\code{course}}{Name of course, eg., "Introduction to Making Things Right"}
  \item{\code{referee}}{Referee, professor, supervisor}
  \item{\code{street}}{Author street}
  \item{\code{zipcity}}{Author zip code and city}
  \item{\code{street}}{Author street}
}

The vignette source shows several of these options in use.
}

\examples{
\dontrun{
rmarkdown::draft("my_report.Rmd", template = "pdf", package = "yart")
rmarkdown::render("my_report.Rmd")
}

}
\references{
JJ Allaire, R Foundation, Hadley Wickham, Journal of Statistical Software, Yihui Xie, Ramnath
Vaidyanathan, Association for Computing Machinery, Carl Boettiger, Elsevier, Karl Broman,
Kirill Mueller, Bastiaan Quast, Randall Pruim, Ben Marwick, Charlotte Wickham, Oliver Keyes
and Miao Yu (2017). rticles: Article Formats for R Markdown. R package version 0.4.1.
\url{https://CRAN.R-project.org/package=rticles}

JJ Allaire, Joe Cheng, Yihui Xie, Jonathan McPherson, Winston Chang, Jeff Allen, Hadley
Wickham, Aron Atkins, Rob Hyndman and Ruben Arslan (2017). rmarkdown: Dynamic Documents for R.
R package version 1.6. \url{https://CRAN.R-project.org/package=rmarkdown}

Yihui Xie (2017). knitr: A General-Purpose Package for Dynamic Report Generation in R. R
package version 1.17.

Aaron Wolen (2017). pandoc-letter. GitHub Repository. \url{https://github.com/aaronwolen/pandoc-letter}.

 Dirk Eddelbuettel and Aaron Wolen (2017). linl: 'linl' is not 'Letter'. R package version 0.0.2. \url{https://CRAN.R-project.org/package=linl}
}
\seealso{
\code{\link[pinp]{pinp}}
\code{\link[linl]{linl}}
\code{\link[tint]{tint}}
}
